requirements.txt
django==4.2.10
google-generativeai>=0.8.3
python-dotenv